Build also Gazebo
Experimental build that also includes Gazebo.

On the top of the usual `vcpkg-robotology.zip` archive, this  PR also generates the archive `vcpkg-robotology-with-gazebo.zip`,
that is like the usual vcpkg-robotology.zip archive, but that also contains a `C:\robotology\gazebo\install` directory in which gazebo 11
and its dependencies are installed.

As Gazebo in this case it is  not installed via vcpkg, it is not possible to find it via the official vcpkg's integration via the
CMAKE_TOOLCHAIN_FILE, so also a set of scripts is provided to set for a given terminal (setup-deps.bat and setup-deps.sh) or for the
system (addPathsToUserEnvVariables-deps.ps1 and removePathsFromUserEnvVariables-deps.ps1) the necessary enviroment variables to
find the dependencies and run Gazebo. The scripts have a behaviour similar to the one already created and installed by the robotology-superbuild.

The specific version of Gazebo installed by the scripts are contained in the gazebo-repos.yaml file. They do refer to specific commit due to the fact
that the Gazebo windows support requires some of the latest fix, but they will be updated to tags once a release will be made including the required tags.